Eliminando producto y mostrando alertas

+++ src/components/Producto.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import {Link} from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 //Redux
 import { useDispatch } from 'react-redux';
@@ -15,9 +16,23 @@ const Producto = ({producto}) => {
     const confirmarEliminarProducto = id =>{
 
         //Preguntar al usuario
+        Swal.fire({
+            title: '¿Estás Seguro?',
+            text: "Esta opción es irreversible!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, eliminar!',
+            cancelButtonText: 'Cancelar'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                //Pasarlo al action
+                dispatch(eliminarProductoAction(id));
+                
+            }
+          })
         
-        //Pasarlo al action
-        dispatch(eliminarProductoAction(id));
     } 
 
     return (
